backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==5.0.0
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.10
openai==1.57.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
werkzeug==3.1.3
gunicorn==23.0.0

